requirements.txt
alembic==0.9.3.dev0
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.4.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==6.7
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.1.4
django-allauth==0.39.1
django-ckeditor==5.7.0
django-constance==2.4.0
django-cors-headers==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-markdown2==0.3.1
django-markdownx==2.0.28
django-picklefield==2.0
django-rest-auth==0.9.3
django-rest-passwordreset==1.0.0
django-smart-selects==1.5.4
django-storages==1.7.1
djangorestframework==3.9.0
feedparser==5.2.1
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Gravatar==0.4.2
Flask-HTMLmin==1.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
gunicorn==19.9.0
html5lib==0.999999999
htmlmin==0.1.12
httplib2==0.9.2
idna==2.8
isort==4.3.18
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.0
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1
markdown2==2.3.7
MarkupSafe==1.0
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.16
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.0.1
olefile==0.45.1
paramiko==2.0.0
passlib==1.7.1
pexpect==4.2.1
Pillow==6.0.0
protobuf==3.7.1
psutil==5.4.2
psycopg2==2.8.2
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.2.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
sendgrid==3.6.5
sendgrid-django==4.2.0
simplejson==3.13.2
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
SQLAlchemy==1.1.11
sqlparse==0.2.4
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
typed-ast==1.3.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24.3
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.1
WTForms==2.1
xkit==0.0.0
zope.interface==4.3.2
